change type of weiAmount from Int to Long so that it can accept large amount of ether

+++ cross-chain-atomic-swap-cordapp/src/main/kotlin/jp/co/layerx/cordage/crosschainatomicswap/flow/ProposeAtomicSwapFlow.kt
@@ -19,7 +19,7 @@ import net.corda.core.utilities.ProgressTracker.Step
 class ProposeAtomicSwapFlow(
     private val securityLinearId: String,
     private val securityAmount: Int,
-    private val weiAmount: Int,
+    private val weiAmount: Long,
     private val swapId: String,
     private val acceptor: Party,
     private val FromEthereumAddress: String,
